requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.1
django-browser-reload==1.18.0
django-class-based-auth-views-jp==0.3.1
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
fonttools==4.58.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mysqlclient==2.2.7
networkx==3.5
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
watchfiles==1.0.3
websockets==14.1
whitenoise==6.9.0